# back/build/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Mindi_API
  description: |
    # Mindi의 API 명세서
servers:
  - url: 'http://localhost:5001/api/'
...
paths:
  /diaries:
    post:
      tags:
        - diaries
      summary: 다이어리를 생성합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - diary
                - feeling
                - sentiment
              properties:
                diary:
                  type: string
                  description: 일기 내용이 들어갑니다.
                feeling:
                  type: string
                  description: 감정 내용이 들어갑니다.
                sentiment:
                  type: object
                  description: 감정 분석 결과가 들어갑니다.
                background:
                  type: file
                  format: binary
                  description: (Optinal)사진을 첨부할 수 있습니다. 5MB 이하의 파일만 업로드 가능합니다.
      responses:
        '201':
          description: A result object.
          content:
            application/json:
              schema:
                $ref: '#/paths/~1diaries/put/responses/200/content/application~1json/schema'
              examples:
                newDiaryWithImage:
                  $ref: '#/paths/~1diaries~1%7Bid%7D/get/responses/200/content/application~1json/examples/diary'
                newDiaryWithoutImage:
                  value:
                    success: true
                    diary:
                      userId: 3e37d05b9fc9d01c9cdf5ac3
                      diary: example
                      feeling: example
                      sentiment:
                        fear: 2
                        surprised: 0
                        anger: 0
                        sadness: 0
                        happiness: 1
                        aversion: 0
                      createdDate: 2022-6-14
                      _id: 62a7ed0f291754e07d151aac
                      createdAt: '2022-06-14T02:06:07.080Z'
                      updatedAt: '2022-06-14T02:06:07.080Z'
                      __v: 0
    put:
      tags:
        - diaries
      summary: 다이어리를 업데이트 합니다.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - _id
                - diary
                - feeling
                - sentiment
                - createdDate
              properties:
                _id:
                  type: string
                  description: 다이어리의 object ID가 들어갑니다. 수정이 불가능한 속성입니다.
                diary:
                  type: string
                  description: 일기 내용이 들어갑니다.
                feeling:
                  type: string
                  description: 감정 내용이 들어갑니다.
                sentiment:
                  type: object
                  description: 감정 분석 결과가 들어갑니다.
                createdDate:
                  type: string
                  description: 일기 작성 날짜가 들어갑니다. 수정이 불가능한 속성입니다.
                imageFileName:
                  type: string
                  description: (Optional)사진이 있는 일기라면 사진 정보도 필요합니다.
                background:
                  type: file
                  format: binary
                  description: (Optinal)사진을 첨부할 수 있습니다. 5MB 이하의 파일만 업로드 가능합니다.
      responses:
        '200':
          description: A result object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  diary:
                    type: object
                    required:
                      - _id
                      - userId
                      - diary
                      - feeling
                      - sentiment
                      - createdDate
                    properties:
                      _id:
                        type: string
                      userId:
                        type: string
                      diary:
                        type: string
                      feeling:
                        type: string
                      sentiment:
                        type: object
                      createdDate:
                        type: string
                        description: 오늘(일기 작성) 날짜가 "년도-월-일" 형식으로 저장됩니다.
                      imageFileName:
                        type: string
                      imageFilePath:
                        type: string
              examples:
                updatedDiaryWithImage:
                  $ref: '#/paths/~1diaries~1%7Bid%7D/get/responses/200/content/application~1json/examples/diary'
                updatedDiaryWithoutImage:
                  $ref: '#/paths/~1diaries/post/responses/201/content/application~1json/examples/newDiaryWithoutImage'
    delete:
      tags:
        - diaries
      summary: 다이어리를 삭제합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - _id
              properties:
                _id:
                  type: string
                  description: 다이어리의 object ID가 들어갑니다. 수정이 불가능한 속성입니다.
                imageFileName:
                  type: string
                  description: (Optional)사진이 있는 일기라면 사진 정보도 필요합니다.
            examples:
              diaryToRemove:
                value:
                  _id: 62a7ed0f291754e07d151aac
                  imageFileName: background_1655172367029_test.jpg
                  imageFilePath: 'http://localhost:5001/images/background_1655172367029_test.jpg'
              diaryWithNoImage:
                value:
                  _id: 62a7ed0f291754e07d151aac
      responses:
        '204':
          description: No content
  '/diaries?date={createdDate}':
    get:
      tags:
        - diaries
      summary: 특정 날짜의 다이어리 리스트를 반환합니다.
      parameters:
        - in: query
          name: date
          schema:
            type: string
            example: '2022-06-16'
      responses:
        '200':
          description: A result object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  diaries:
                    type: array
                    items:
                      types: object
                      required:
                        - _id
                        - userId
                        - diary
                        - feeling
                        - sentiment
                        - createdDate
                      properties:
                        _id:
                          type: string
                        userId:
                          type: string
                        diary:
                          type: string
                        feeling:
                          type: string
                        sentiment:
                          type: object
                        createdDate:
                          type: string
                        imageFileName:
                          type: string
                        imageFilePath:
                          type: string
              examples:
                diaryList:
                  value:
                    success: true
                    diaries:
                      - _id: 62aaf383abc6f9d6325f06e4
                        userId: 62a76c45f0ed80e0f9e765fc
                        diary: Consequuntur praesentium nesciunt sit maiores. Et eos eum adipisci et quis nam repudiandae soluta. Ut provident natus maiores quasi.
                        feeling: Minus nostrum soluta ratione sed saepe dolorum corrupti.
                        sentiment:
                          fear: 2
                          surprised: 0
                          anger: 0
                          sadness: 0
                          happiness: 1
                          aversion: 0
                        createdDate: '2022-06-16'
                        createdAt: '2022-06-16T09:10:27.553Z'
                        updatedAt: '2022-06-16T09:10:27.706Z'
                        __v: 0
                      - _id: 62aaec9597623ebf98269f97
                        userId: 62a76c45f0ed80e0f9e765fc
                        diary: Dolor nihil magnam distinctio dolores molestias quidem harum et. Voluptatem sapiente totam. A animi asperiores aut deleniti cumque odit neque ipsum labore. Adipisci vel doloremque. Amet unde repellendus dicta voluptas provident similique at suscipit.
                        feeling: Ab fugiat tempora.
                        sentiment:
                          fear: 2
                          surprised: 0
                          anger: 0
                          sadness: 0
                          happiness: 1
                          aversion: 0
                        createdDate: '2022-06-16'
                        createdAt: '2022-06-16T08:40:53.671Z'
                        updatedAt: '2022-06-16T08:40:53.671Z'
                        __v: 0
  '/diaries/{id}':
    get:
      tags:
        - diaries
      summary: 특정 아이디의 다이어리 정보를 반환합니다.
      parameters:
        - in: path
          name: id
          schema:
            type: string
            example: 3e37d05b9fc9d01c9cdf5ac3
          description: The diary ID
      responses:
        '200':
          description: A result object.
          content:
            application/json:
              schema:
                $ref: '#/paths/~1diaries/put/responses/200/content/application~1json/schema'
              examples:
                diary:
                  value:
                    success: true
                    diary:
                      userId: 3e37d05b9fc9d01c9cdf5ac3
                      diary: example
                      feeling: example
                      sentiment:
                        fear: 2
                        surprised: 0
                        anger: 0
                        sadness: 0
                        happiness: 1
                        aversion: 0
                      createdDate: 2022-6-14
                      imageFileName: background_1655172367029_test.jpg
                      imageFilePath: 'http://localhost:5001/images/background_1655172367029_test.jpg'
                      _id: 62a7ed0f291754e07d151aac
                      createdAt: '2022-06-14T02:06:07.080Z'
                      updatedAt: '2022-06-14T02:06:07.080Z'
                      __v: 0
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer